net-test: transperf: in gen_transperf_pages(), show retx as a percent
For the page that summarizes the result of all experiments, show
retransmit rates as percent (rather than fraction).

This makes the table consistent with the summary in each experiment
dashboard, which already presented the retransmit rate as a percent.
See SummaryPageGenerator.end(), which used a format of:
  "%.2f%%"
and a value of:
  self._metrics['retx'].get(port)*100

In addition to consistency, this shows more digits of
precision. Previous to this change, loss rates of 0.07% were rounded
up to show as 0.001 and loss rates of 0.03% rounded down to show as
0.000. This extra precision is important for Reno/CUBIC and
coexistence with those algorithms, which have very different
throughputs depending on small differences in small loss rates.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: I0e889a653e1f7647fbe9537f0e66799ebf28df97